HircFramework

Documentation

type InternalChan = Chan (Target, String)

Used for internal commnands like Logging (Log). Don't use this directly, pass it as an argument.

type ExternalChan = Chan Event

Used for external commands, like sending a private message (PvtMsg). Use helper functions instead of using this directly, for example privateMessage.

data Event

Events from the server. Patternmatch this in your BotThread.

Constructors

EReplyYourHost ServerName Version

EReplyCreated Date

EReplyMyInfo ServerName Version UserModes ChannelModes

EPing Message

EPrivateMessage Nick Message

EChannelMessage Channel Nick Message

EPart Channel Nick Message

EJoin Channel Nick

EInvite Channel Nick

EKick Channel Nick Nick Message

ETopic Channel Nick Message

EPrivateNotice Nick Message

EChannelNotice Channel Nick Message

EUserMode Nick Nick Modes

EChannelMode Channel Nick Modes

EChannelUserMode Channel Nick [(Prefix, Nick)]

ENames Channel [(Prefix, Nick)]

EMotd [String]

ENick Nick Nick

EAway

type HircHandle = (Handle, InternalChan, ExternalChan)

Wraps three things required all over Hirc to simplify usage

type BotFunc = HircHandle -> IO ()

Will be called when the bot has connected to the server

startBot :: ServerInfo -> BotInfo -> BotFunc -> IO ()

This function starts the bot

logMessage :: HircHandle -> String -> IO ()

privateMessage :: HircHandle -> String -> String -> IO ()

channelMessage :: HircHandle -> String -> String -> IO ()

privateNotice :: HircHandle -> String -> String -> IO ()

channelNotice :: HircHandle -> String -> String -> IO ()

join :: HircHandle -> String -> IO ()

nick :: HircHandle -> String -> IO ()

user :: HircHandle -> String -> String -> IO ()

pong :: HircHandle -> String -> IO ()

whois :: HircHandle -> String -> IO ()

whowas :: HircHandle -> String -> IO ()

kick :: HircHandle -> String -> String -> String -> IO ()

away :: HircHandle -> String -> IO ()

unAway :: HircHandle -> IO ()

rehash :: HircHandle -> IO ()

users :: HircHandle -> IO ()

part :: HircHandle -> String -> String -> IO ()

userMode :: HircHandle -> String -> String -> IO ()

userChannelMode :: HircHandle -> String -> String -> String -> IO ()

channelMode :: HircHandle -> String -> String -> IO ()

getTopic :: HircHandle -> String -> IO ()

setTopic :: HircHandle -> String -> String -> IO ()

raw :: HircHandle -> String -> IO ()

names :: HircHandle -> String -> IO ()

quit :: HircHandle -> String -> IO ()

invite :: HircHandle -> String -> String -> IO ()